+++ src/layers/heatmap-layer.ts
