+++ src/layers/point-cloud-layer.ts
