+++ src/layers/text-layer.ts
@@ -16,15 +16,15 @@ import {
   expandArrayToCoords,
   extractAccessorsFromProps,
   getGeometryVector,
-} from "./utils.js";
+} from "../utils.js";
 import {
   GeoArrowExtraPickingProps,
   computeChunkOffsets,
   getPickingInfo,
-} from "./picking.js";
-import { ColorAccessor, FloatAccessor, GeoArrowPickingInfo } from "./types.js";
-import { EXTENSION_NAME } from "./constants.js";
-import { validateAccessors } from "./validate.js";
+} from "../utils/picking.js";
+import { ColorAccessor, FloatAccessor, GeoArrowPickingInfo } from "../types.js";
+import { EXTENSION_NAME } from "../constants.js";
+import { validateAccessors } from "../utils/validate.js";
 
 /** All properties supported by GeoArrowTextLayer */
 export type GeoArrowTextLayerProps = Omit<
